moved keyMgmt and tiered dependencies
