+++ it/src/test/scala/co/topl/it/util/KeyCreator.scala
@@ -2,7 +2,8 @@ package co.topl.it.util
 
 import co.topl.attestation.Address
 import co.topl.attestation.AddressEncoder.NetworkPrefix
-import co.topl.crypto.{KeyRing, KeyfileCurve25519, PrivateKeyCurve25519}
+import co.topl.attestation.keyManagement.{KeyfileCurve25519, PrivateKeyCurve25519}
+import co.topl.crypto.KeyfileCurve25519
 import co.topl.settings.NetworkType._
 
 import scala.util.Try
