+++ src/main/scala/co/topl/attestation/Proposition.scala
@@ -2,8 +2,8 @@ package co.topl.attestation
 
 import co.topl.attestation.AddressEncoder.NetworkPrefix
 import co.topl.attestation.Evidence.{EvidenceContent, EvidenceTypePrefix}
+import co.topl.attestation.keyManagement.{PrivateKeyCurve25519, Secret}
 import co.topl.attestation.serialization.PropositionSerializer
-import co.topl.crypto.{PrivateKeyCurve25519, Secret}
 import co.topl.utils.serialization.{BifrostSerializer, BytesSerializable}
 import co.topl.utils.{Identifiable, Identifier}
 import com.google.common.primitives.Ints
