+++ src/main/scala/co/topl/attestation/keyManagement/KeyfileCurve25519.scala
@@ -1,4 +1,4 @@
-package co.topl.crypto
+package co.topl.attestation.keyManagement
 
 import java.nio.charset.StandardCharsets
