+++ src/main/scala/co/topl/attestation/keyManagement/PrivateKeyCurve25519.scala
@@ -1,4 +1,4 @@
-package co.topl.crypto
+package co.topl.attestation.keyManagement
 
 import co.topl.attestation.{PublicKeyPropositionCurve25519, SignatureCurve25519}
 import co.topl.utils.serialization.{BifrostSerializer, Reader, Writer}
