+++ src/main/scala/co/topl/attestation/keyManagement/SecretGenerator.scala
@@ -1,4 +1,4 @@
-package co.topl.crypto
+package co.topl.attestation.keyManagement
 
 /**
  * SecretGenerator is a type-class that describes a class ability to generate a secret
