+++ src/main/scala/co/topl/consensus/genesis/GenesisProvider.scala
@@ -1,8 +1,8 @@
 package co.topl.consensus.genesis
 
 import co.topl.attestation.AddressEncoder.NetworkPrefix
+import co.topl.attestation.keyManagement.PrivateKeyCurve25519
 import co.topl.consensus.Forger.ChainParams
-import co.topl.crypto.PrivateKeyCurve25519
 import co.topl.modifier.ModifierId
 import co.topl.modifier.block.Block
 import co.topl.modifier.block.PersistentNodeViewModifier.PNVMVersion
