+++ src/test/scala/co/topl/serialization/JsonTests.scala
@@ -1,7 +1,7 @@
 package co.topl.serialization
 
+import co.topl.attestation.keyManagement.KeyfileCurve25519
 import co.topl.attestation.{Address, Evidence, PublicKeyPropositionCurve25519, SignatureCurve25519}
-import co.topl.crypto.KeyfileCurve25519
 import co.topl.modifier.ModifierId
 import co.topl.modifier.block.{Block, BlockBody, BlockHeader, BloomFilter}
 import co.topl.modifier.transaction.Transaction
