+++ src/test/scala/co/topl/utils/GenesisGenerators.scala
@@ -1,7 +1,8 @@
 package co.topl.utils
 
+import co.topl.attestation.keyManagement.{KeyRing, KeyfileCurve25519, PrivateKeyCurve25519}
 import co.topl.consensus.genesis.PrivateGenesis
-import co.topl.crypto.{KeyRing, KeyfileCurve25519, PrivateKeyCurve25519}
+import co.topl.crypto.KeyfileCurve25519
 import co.topl.modifier.ModifierId
 import co.topl.modifier.block.Block
 import co.topl.nodeView.history.History
