+++ src/test/scala/co/topl/utils/ValidGenerators.scala
@@ -2,8 +2,9 @@ package co.topl.utils
 
 import co.topl.attestation.PublicKeyPropositionCurve25519
 import co.topl.attestation.PublicKeyPropositionCurve25519.evProducer
+import co.topl.attestation.keyManagement.{KeyRing, KeyfileCurve25519, PrivateKeyCurve25519}
 import co.topl.consensus.genesis.PrivateGenesis
-import co.topl.crypto.{KeyRing, KeyfileCurve25519, PrivateKeyCurve25519}
+import co.topl.crypto.KeyfileCurve25519
 import co.topl.modifier.ModifierId
 import co.topl.modifier.block.Block
 import co.topl.modifier.box.{AssetCode, AssetValue, SecurityRoot}
